requirements.txt
asgiref==3.9.1
black==25.11.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Django==5.2.8
djangorestframework==3.16.1
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mypy_extensions==1.1.0
mysqlclient==2.2.7
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pytokens==0.3.0
sqlparse==0.5.3
tzdata==2025.2
Werkzeug==3.1.3
